+++ resources/lang/en/validation.php
@@ -60,4 +60,8 @@ return [
     'date_time_iso_8601' => 'The :attribute is not a valid ISO 8601 date and time.',
 
     'disallowed_parameter' => 'Parameter :name is not allowed.',
+
+    'has_one' => 'The :attribute field must be a to-one relationship containing :types resources.',
+
+    'has_many' => 'The :attribute field must be a to-many relationship containing :types resources.',
 ];