+++ src/Api/Repository.php
@@ -24,6 +24,7 @@ use CloudCreativity\LaravelJsonApi\Exceptions\RuntimeException;
 use CloudCreativity\LaravelJsonApi\Factories\Factory;
 use CloudCreativity\LaravelJsonApi\Resolver\AggregateResolver;
 use Illuminate\Contracts\Config\Repository as Config;
+use Illuminate\Support\Arr;
 
 /**
  * Class Repository
@@ -87,7 +88,9 @@ class Repository
             $config['use-eloquent'],
             $config['supported-ext'],
             $config['errors'],
-            $config['providers'] ?? []
+            $config['providers'] ?? [],
+            $config['controllers']['connection'] ?? null,
+            $config['controllers']['transactions'] ?? true
         );
 
         /** Attach resource providers to the API. */
@@ -132,6 +135,7 @@ class Repository
         $config['resources'] = $this->normalizeResources($config['resources'] ?? [], $config);
         $config['url'] = $this->normalizeUrl($config['url'] ?? [], $host);
         $config['errors'] = array_replace($this->defaultErrors(), $config['errors'] ?? []);
+        $config['controllers'] = $config['controllers'] ?? [];
 
         return $config;
     }
@@ -163,7 +167,7 @@ class Repository
      */
     private function normalizeUrl(array $url, $host = null)
     {
-        $prependHost = false !== array_get($url, 'host');
+        $prependHost = false !== Arr::get($url, 'host');
 
         if ($host) {
             $url['host'] = $host;
@@ -173,8 +177,8 @@ class Repository
 
         return [
             'host' => $prependHost ? (string) $url['host'] : '',
-            'namespace' => (string) array_get($url, 'namespace'),
-            'name' => (string) array_get($url, 'name'),
+            'namespace' => (string) Arr::get($url, 'namespace'),
+            'name' => (string) Arr::get($url, 'name'),
         ];
     }